+++ frontend/src/app/App.tsx
@@ -2,7 +2,15 @@ import { Admin } from "react-admin";
 
 import { Dashboard } from "../features/dashboard/Dashboard";
 import { authProvider } from "../services/authProvider";
+import { dataProvider } from "../services/dataProvider";
 
 export function App() {
-  return <Admin authProvider={authProvider} dashboard={Dashboard} disableTelemetry />;
+  return (
+    <Admin
+      authProvider={authProvider}
+      dashboard={Dashboard}
+      dataProvider={dataProvider}
+      disableTelemetry
+    />
+  );
 }